# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API server for managing Kibaship Operator resources",
    "title": "Kibaship Operator API",
    "termsOfService": "https://github.com/kibamail/kibaship",
    "contact": {
      "name": "API Support",
      "email": "support@kibamail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/v1",
  "paths": {
...
    "/v1/applications/{uuid}/domains": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Create a new domain for an application",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["application-domains"],
        "summary": "Create a new application domain",
        "parameters": [
          {
            "type": "string",
            "description": "Application UUID or slug (8-character identifier)",
            "name": "uuid",
            "in": "path",
            "required": true
          },
          {
            "description": "Application domain creation data",
            "name": "domain",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.ApplicationDomainCreateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Application domain created successfully",
            "schema": {
              "$ref": "#/definitions/models.ApplicationDomainResponse"
            }
          },
          "400": {
            "description": "Validation errors in request data",
            "schema": {
              "$ref": "#/definitions/models.ValidationErrors"
            }
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Application not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/v1/domains/{uuid}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Retrieve an application domain by its unique UUID or slug identifier",
        "produces": ["application/json"],
        "tags": ["application-domains"],
        "summary": "Get application domain by UUID",
        "parameters": [
          {
            "type": "string",
            "description": "Application domain UUID or slug (8-character identifier)",
            "name": "uuid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Application domain details",
            "schema": {
              "$ref": "#/definitions/models.ApplicationDomainResponse"
            }
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Application domain not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Delete an application domain by its unique UUID or slug identifier",
        "tags": ["application-domains"],
        "summary": "Delete application domain by UUID",
        "parameters": [
          {
            "type": "string",
            "description": "Application domain UUID or slug (8-character identifier)",
            "name": "uuid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Application domain deleted successfully"
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Application domain not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "auth.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Unauthorized"
        },
        "message": {
          "type": "string",
          "example": "Invalid API key"
        }
      }
    },
...
    "models.ApplicationDomainCreateRequest": {
      "type": "object",
      "required": ["applicationSlug", "domain", "port"],
      "properties": {
        "applicationSlug": {
          "type": "string",
          "example": "abc123de"
        },
        "default": {
          "type": "boolean",
          "example": false
        },
        "domain": {
          "type": "string",
          "example": "my-app.example.com"
        },
        "port": {
          "type": "integer",
          "maximum": 65535,
          "minimum": 1,
          "example": 3000
        },
        "tlsEnabled": {
          "type": "boolean",
          "example": true
        },
        "type": {
          "allOf": [
            {
              "$ref": "#/definitions/models.ApplicationDomainType"
            }
          ],
          "example": "custom"
        }
      }
    },
    "models.ApplicationDomainPhase": {
      "type": "string",
      "enum": ["Pending", "Ready", "Failed"],
      "x-enum-varnames": [
        "ApplicationDomainPhasePending",
        "ApplicationDomainPhaseReady",
        "ApplicationDomainPhaseFailed"
      ]
    },
    "models.ApplicationDomainResponse": {
      "type": "object",
      "properties": {
        "applicationSlug": {
          "type": "string",
          "example": "abc123de"
        },
        "applicationUuid": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440001"
        },
        "certificateReady": {
          "type": "boolean",
          "example": false
        },
        "createdAt": {
          "type": "string",
          "example": "2023-01-01T12:00:00Z"
        },
        "default": {
          "type": "boolean",
          "example": false
        },
        "dnsConfigured": {
          "type": "boolean",
          "example": false
        },
        "domain": {
          "type": "string",
          "example": "my-app.example.com"
        },
        "ingressReady": {
          "type": "boolean",
          "example": false
        },
        "phase": {
          "allOf": [
            {
              "$ref": "#/definitions/models.ApplicationDomainPhase"
            }
          ],
          "example": "Pending"
        },
        "port": {
          "type": "integer",
          "example": 3000
        },
        "projectUuid": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440002"
        },
        "slug": {
          "type": "string",
          "example": "def456gh"
        },
        "tlsEnabled": {
          "type": "boolean",
          "example": true
        },
        "type": {
          "allOf": [
            {
              "$ref": "#/definitions/models.ApplicationDomainType"
            }
          ],
          "example": "custom"
        },
        "updatedAt": {
          "type": "string",
          "example": "2023-01-01T12:00:00Z"
        },
        "uuid": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        }
      }
    },
    "models.ApplicationDomainType": {
      "type": "string",
      "enum": ["default", "custom"],
      "x-enum-varnames": [
        "ApplicationDomainTypeDefault",
        "ApplicationDomainTypeCustom"
      ]
    },
...
    "models.ValidationError": {
      "type": "object",
      "properties": {
        "field": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "models.ValidationErrors": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.ValidationError"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}